refactor(#287): 매칭검사 저장 로직 수정 및 마이페이지 내 특성 로직 분리 2
